bot/pyproject.toml
[tool.poetry]
name = "bot"
version = "0.1.0"
description = ""
authors = ["Eckorezze <alexandrsoloby@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aiogram = ">=2.25.1,<2.26.0"
segno = ">=1.5.2,<1.6.0"
requests = ">=2.31.0,<2.32.0"
flask = ">=2.2.5,<2.3.0"
pip = ">=23.1.2,<23.2.0"
colorama = ">=0.4.6,<0.5.0"
pytz = ">=2023.3,<2024.0"
idna = ">=3.4,<4.0"
multidict = ">=6.0.4,<6.1.0"
attrs = ">=23.1.0,<23.2.0"
markupsafe = ">=2.1.3,<2.2.0"
werkzeug = ">=2.3.6,<2.4.0"
click = ">=8.1.6,<8.2.0"
jinja2 = ">=3.1.2,<3.2.0"
itsdangerous = ">=2.1.2,<2.2.0"
numpy = ">=1.25.2,<1.26.0"
setuptools = ">=65.5.0,<65.6.0"
cffi = ">=1.15.1,<1.16.0"
six = ">=1.16.0,<1.17.0"
urllib3 = ">=1.26.15,<1.27.0"
certifi = ">=2022.12.7,<2022.13.0"
pandas = ">=2.0.3,<2.1.0"
openpyxl = ">=3.1.2,<3.2.0"
python-dotenv = ">=1.0.0,<1.1.0"
pillow = ">=10.0.1,<10.1.0"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
bot = "src.bot.main:main"